+++ tests/Type/ExternalReplyInfoTest.php
@@ -182,6 +182,7 @@ final class ExternalReplyInfoTest extends TestCase
                 'is_anonymous' => false,
                 'type' => 'regular',
                 'allows_multiple_answers' => true,
+                'allows_revoting' => false,
             ],
             'venue' => [
                 'location' => [