+++ tests/Type/PollTest.php
@@ -19,7 +19,7 @@ final class PollTest extends TestCase
 {
     public function testBase(): void
     {
-        $option = new PollOption('One', 12);
+        $option = new PollOption('pid1', 'One', 12);
         $poll = new Poll(
             '12',
             'Why?',
@@ -29,6 +29,7 @@ final class PollTest extends TestCase
             false,
             'regular',
             true,
+            false,
         );
 
         assertSame('12', $poll->id);
@@ -39,11 +40,14 @@ final class PollTest extends TestCase
         assertFalse($poll->isAnonymous);
         assertSame('regular', $poll->type);
         assertTrue($poll->allowsMultipleAnswers);
-        assertNull($poll->correctOptionId);
+        assertFalse($poll->allowsRevoting);
+        assertNull($poll->correctOptionIds);
         assertNull($poll->explanation);
         assertNull($poll->explanationEntities);
         assertNull($poll->openPeriod);
         assertNull($poll->closeDate);
+        assertNull($poll->description);
+        assertNull($poll->descriptionEntities);
     }
 
     public function testFromTelegramResult(): void
@@ -52,13 +56,14 @@ final class PollTest extends TestCase
             'id' => '12',
             'question' => 'Why?',
             'options' => [
-                ['text' => 'One', 'voter_count' => 12],
+                ['persistent_id' => 'pid1', 'text' => 'One', 'voter_count' => 12],
             ],
             'total_voter_count' => 42,
             'is_closed' => true,
             'is_anonymous' => false,
             'type' => 'regular',
             'allows_multiple_answers' => true,
+            'allows_revoting' => true,
             'question_entities' => [
                 [
                     'offset' => 0,
@@ -66,7 +71,7 @@ final class PollTest extends TestCase
                     'type' => 'bold',
                 ],
             ],
-            'correct_option_id' => 23,
+            'correct_option_ids' => [0, 2],
             'explanation' => 'Because',
             'explanation_entities' => [
                 [
@@ -77,6 +82,14 @@ final class PollTest extends TestCase
             ],
             'open_period' => 123,
             'close_date' => 456,
+            'description' => 'Poll description',
+            'description_entities' => [
+                [
+                    'offset' => 0,
+                    'length' => 4,
+                    'type' => 'bold',
+                ],
+            ],
         ], null, Poll::class);
 
         assertSame('12', $poll->id);
@@ -90,11 +103,12 @@ final class PollTest extends TestCase
         assertFalse($poll->isAnonymous);
         assertSame('regular', $poll->type);
         assertTrue($poll->allowsMultipleAnswers);
+        assertTrue($poll->allowsRevoting);
 
         assertCount(1, $poll->questionEntities);
         assertSame(35, $poll->questionEntities[0]->length);
 
-        assertSame(23, $poll->correctOptionId);
+        assertSame([0, 2], $poll->correctOptionIds);
         assertSame('Because', $poll->explanation);
 
         assertCount(1, $poll->explanationEntities);
@@ -102,5 +116,8 @@ final class PollTest extends TestCase
 
         assertSame(123, $poll->openPeriod);
         assertSame(456, $poll->closeDate);
+        assertSame('Poll description', $poll->description);
+        assertCount(1, $poll->descriptionEntities);
+        assertSame(4, $poll->descriptionEntities[0]->length);
     }
 }